[webapp] add shared Telegram theme module
